Add solutions for Day 11 Part 1 & Part 2*
Solutions are not implemented separately. The input effectively
changes for part 2, and I didn't feel like coding in the additional
input manually. The program speed is improved by adjusting a cuttoff
criteria, but the current determination of that criteria was just
trial an error and isn't programatic (the values that work are
different for the part 1 and part 2 inputs).