Improve speed of Part 2 solution
Modified is_safe() to return the starting index of the first unsafe
interval in a report. Then dampened() simply checks the index 0 of
the report, the reported starting index, and the next index to see if
removing any one of those makes the report safe. This cuts the run-time
for part 2 in half, roughly.